Issue 85
Added animation to delete Rightmost for number to fly off canvas to look more appealing.
Switched Find to Search
Added a circle if value was found using Search with the help of SimpleSorting.py & Array,py.

 #NOTE WHEN 86 WILL MERGE
We will need to revise so that the Search is compatible with the other methods.