Fix deletions of canvas items in siftDown
The primary problem was due to the assignment of one drawable record
to multiple cells in the self.list array.  When two cells point to the
same drawable record, operations on one cell like deleting the
display_shape operate on both cells because they share the record.
So lines 233 and 261 had to be removed.  Instead of copying the entire
record, individual fields of the record are copied as needed.

In the sift down operation, values are copied up the heap (from the
maximum child, maxi, to its parent, i).  I changed the operations
inside the sift down loop to copy from maxi to i.

I also made two other small changes:

 * In line 207, the childIndex starts off as the left child, not the parent

 * In line 488 inside the display() method, the fill color has to be
   extracted from the last item in the tuple returned by itemconfigure