Update Heap.py
Took out the self.size attribute and replaced it with self.nItems and len(self.list) throughout the code. Still some bugs with it. Also fixed the removeMax() to make the top tile display itself on the bottom of the screen instead of rising to the top off the screen